# lib/api-zod/src/generated/types/createBountyBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Team Fate Shiny Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateBountyBody {
  title: string;
  description: string;
  imageUrl?: string | null;
  month: string;
  isActive: boolean;
  points?: number | null;
}