# lib/api-zod/src/generated/types/createShinyBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Team Fate Shiny Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateShinyBody {
  pokemonId: number;
  pokemonName: string;
  pokemonSpriteUrl: string;
  memberId: number;
  shinyTypeId?: number | null;
  caughtAt: string;
  catchMethod?: string | null;
  encounterNumber?: number | null;
  location?: string | null;
  notes?: string | null;
  isAlpha: boolean;
  isSecret: boolean;
}