# lib/api-zod/src/generated/types/nextEvent.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Team Fate Shiny Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface NextEvent {
  id: number;
  title: string;
  description: string;
  imageUrl?: string | null;
  externalUrl?: string | null;
  eventDate?: string | null;
}