# lib/api-zod/src/generated/types/uploadResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Team Fate Shiny Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface UploadResult {
  url: string;
  filename: string;
}